new versionCVMBogota

+++ quake/forward/psolve.c
@@ -343,7 +343,7 @@ static struct Global_t {
 
 // Global variables Bogota
 static double ** DataFilesBogota; //array that stores planes characteristics
-static double ** PlanesBogota;    //array that stores all planes data (long lat elevation deth Vp Vs Rho)
+static float  ** PlanesBogota;    //array that stores all planes data (long lat elevation deth Vp Vs Rho)
 
 
 /* ------------------------------End of declarations------------------------------ */
